# APP&WEB/identity/openapi.yaml
openapi: 3.0.3
info:
  title: ZIS — ZION Identity Service
  version: 3.2.0
  description: |
    ZION Identity Service (auth.zionterranova.com). Issues `zion_session`
    httpOnly JWT cookies scoped to `.zionterranova.com`, manages linked
    chain addresses (zion-l1, EVM, bitcoin), API keys, and proxies
    custodial wallet operations to the L2 multichain service.

    Authentication: `zion_session` cookie (browser) or `Authorization:
    Bearer <jwt>` / API key for programmatic access.
servers:
  - url: https://auth.zionterranova.com
  - url: http://127.0.0.1:8096
    description: Edge local
...
paths:
...
  /api/auth/webauthn/register/options:
    post:
      tags: [webauthn]
      summary: Begin passkey registration (returns WebAuthn creation options)
      responses:
        '200':
          description: '{ ceremonyId, options } — pass options to the browser WebAuthn API'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { description: 'LIMIT — maximum of 10 passkeys per account' }
  /api/auth/webauthn/register/verify:
    post:
      tags: [webauthn]
      summary: Finish passkey registration, store credential
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ceremonyId, response]
              properties:
                ceremonyId: { type: string }
                response: { type: object, description: RegistrationResponseJSON from the browser }
                label: { type: string, maxLength: 64 }
      responses:
        '200': { description: 'Credential stored; returns summary without public key' }
        '400': { description: 'BAD_REQUEST | CEREMONY_INVALID | VERIFY_FAILED' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { description: 'CONFLICT — passkey already registered' }
...
  /api/auth/webauthn/login/verify:
    post:
      tags: [webauthn]
      summary: Finish passkey login, issue zion_session cookie
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ceremonyId, response]
              properties:
                ceremonyId: { type: string }
                response: { type: object, description: AuthenticationResponseJSON from the browser }
      responses:
        '200': { description: 'Session issued; `zion_session` cookie set; returns user object' }
        '400': { description: 'BAD_REQUEST | CEREMONY_INVALID' }
        '401': { description: 'AUTH_FAILED — unknown passkey or verification failed' }
  /api/auth/webauthn/credentials:
    get:
      tags: [webauthn]
      summary: List own passkeys (never includes public keys)
      responses:
        '200': { description: '{ credentials[] }' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/auth/webauthn/credentials/{id}:
    delete:
      tags: [webauthn]
      summary: Delete one of own passkeys
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: '{ "ok": true }' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { description: Not found }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: zion_session
    bearerAuth:
      type: http
      scheme: bearer
...
  responses:
...
    Unauthorized:
      description: Missing or invalid session
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string, example: UNAUTHORIZED }
security:
  - sessionCookie: []
  - bearerAuth: []